+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Include `@deprecated` JSDoc if property is marked as `[Obsolete]` #45
+- Include `@deprecated` JSDoc if property is marked as `[Obsolete]` (#45)
 
 ### Fixed