+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Handle nullability for more types, including `string` and records (#51)
 
+### Changed
+
+- Update xunit and friends
+- Update Microsoft.NET.Test.Sdk to 17.10.0
+
 ## [0.10.0] - 2024-04-21
 
 ### Added